+++ frontend/src/views/index/data/index-survey-anime-data.ts
@@ -1,0 +1,6 @@
+import { AnimeData } from '@/util/data';
+
+export interface IndexSurveyAnimeData {
+  anime: AnimeData;
+  result: number;
+}
